Move materialize out of clone
